Refactor image gallery components into separate widgets
- Moved ImageGallery and ImageTile widgets from image_gallery_screen.dart to new files in lib/screens/gallery/widgets/
